+++ frontend/src/app/game-board/game-board.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 
 import { Player, Transaction } from '../types';
-import { DialogService } from '../dialog-service/dialog.service';
-import { TransactionService } from '../transaction-service/transaction.service';
-import { PlayerService } from '../player-service/player.service';
-import { GameService } from '../game.service';
+import { DialogService } from '../services/dialog-service/dialog.service';
+import { TransactionService } from '../services/transaction.service';
+import { PlayerService } from '../services/player.service';
+import { GameService } from '../services/game.service';
 
 @Component({
   selector: 'app-game-board',
