+++ frontend/src/app/services/dialog-service/log-dialog.html
