+++ frontend/src/app/services/dialog-service/piece-select-dialog.html
